"check_style_score -> 5 of 5 passed. style score: 1/1 "

+++ cspp2-assignments/m2/Assignment-3/Solution.java
@@ -25,7 +25,7 @@ public final class Solution {
      *
      * @return     base power exponent
      */
-    public static int power(final int base, final int exponent) {
+    public static long power(final int base, final int exponent) {
         if (exponent == 0) {
             return 1;
         }
@@ -41,7 +41,7 @@ public final class Solution {
         Scanner s = new Scanner(System.in);
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result = power(base, exponent);
+        long result = power(base, exponent);
         System.out.println(result);
     }
 }
